[#28] Separate parsers and serializers
